[ads] all ads-related files are moved into own directory
