+++ partners_api/ads/ads_utils.cpp
@@ -1,4 +1,4 @@
-#include "partners_api/ads_utils.hpp"
+#include "partners_api/ads/ads_utils.hpp"
 
 #include "coding/string_utf8_multilang.hpp"
