+++ partners_api/ads/ads_utils.hpp
