+++ partners_api/ads/banner.hpp
