+++ partners_api/ads/google_ads.hpp
@@ -1,18 +1,11 @@
 #pragma once
 
-#include "partners_api/ads_base.hpp"
+#include "partners_api/ads/ads_base.hpp"
 
 namespace ads
 {
 // Class which matches feature types and facebook banner ids.
-class FacebookPoi : public PoiContainer
-{
-private:
-  // PoiContainerBase overrides:
-  std::string GetBannerForOtherTypes() const override;
-};
-
-class FacebookSearch : public SearchContainerBase
+class Google : public SearchContainerBase
 {
 public:
   // SearchContainerBase overrides:
