+++ partners_api/ads/rb_ads.cpp
@@ -1,4 +1,4 @@
-#include "partners_api/rb_ads.hpp"
+#include "partners_api/ads/rb_ads.hpp"
 
 namespace
 {
