+++ partners_api/partners_api_tests/ads_engine_tests.cpp
@@ -4,11 +4,11 @@
 #include "indexer/classificator_loader.hpp"
 #include "indexer/feature_data.hpp"
 
-#include "partners_api/ads_engine.hpp"
-#include "partners_api/facebook_ads.hpp"
-#include "partners_api/google_ads.hpp"
-#include "partners_api/mopub_ads.hpp"
-#include "partners_api/rb_ads.hpp"
+#include "partners_api/ads/ads_engine.hpp"
+#include "partners_api/ads/facebook_ads.hpp"
+#include "partners_api/ads/google_ads.hpp"
+#include "partners_api/ads/mopub_ads.hpp"
+#include "partners_api/ads/rb_ads.hpp"
 
 namespace
 {
