+++ partners_api/partners_api_tests/google_tests.cpp
@@ -1,6 +1,6 @@
 #include "testing/testing.hpp"
 
-#include "partners_api/google_ads.hpp"
+#include "partners_api/ads/google_ads.hpp"
 
 UNIT_TEST(Google_BannerInSearch)
 {
